asserts,registry: fixes after review
